Playing with quickcheck. Something isn't working, but it may be a NixOS issue.
